move Statistics class to archive module, avoid cyclic import
also move the tests accordingly.
